researchflow-backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.6

# Task queue
arq>=0.26.0
redis>=5.2.0

# PDF parsing
pymupdf>=1.25.0

# HTTP client
httpx>=0.28.0

# YAML parsing (for frontmatter)
pyyaml>=6.0.2

# Object storage
cos-python-sdk-v5>=1.9.30
oss2>=2.19.0

# LLM clients
anthropic>=0.42.0
openai>=1.57.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.12
